+++ lambdas/postgres-migration-async-operation/iam.tf
@@ -70,7 +70,8 @@ data "aws_iam_policy_document" "postgres_migration_async_operation_policy" {
     actions = [
       "dynamodb:GetItem",
       "dynamodb:Scan",
-      "dynamodb:PutItem"
+      "dynamodb:PutItem",
+      "dynamodb:DeleteItem"
     ]
     resources = [for k, v in var.dynamo_tables : v.arn]
   }